--- bin/next-page-bundlesize.js
@@ -1,6 +1,0 @@
-#!/usr/bin/env node
-
-/* eslint-disable @typescript-eslint/no-var-requires */
-/* eslint-disable import/no-internal-modules */
-const run = require("../pkg").default
-run(process.argv)
