+++ tsconfig.json
@@ -1,9 +1,9 @@
 {
   "compilerOptions": {
-    "module": "esnext",
-    "target": "esnext",
+    "module": "commonjs",
+    "target": "es5",
     "moduleResolution": "node",
-    "lib": ["esnext", "dom"],
+    "lib": ["esnext"],
     "declaration": true,
     "esModuleInterop": true,
     "sourceMap": true
